requirements.txt
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.3.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.8.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distro==1.9.0
et-xmlfile==1.1.0
exceptiongroup==1.2.1
Flask==3.0.3
Flask-Cors==4.0.0
Flask-Mail==0.10.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
importlib_metadata==7.1.0
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
MouseInfo==0.1.3
multidict==6.0.5
mysql-connector-python==8.4.0
numpy==1.26.4
openai==1.33.0
openpyxl==3.1.2
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.2
pefile==2023.2.7
pillow==10.3.0
PyAutoGUI==0.9.54
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
pygame==2.5.2
PyGetWindow==0.0.9
pyinstaller==6.9.0
pyinstaller-hooks-contrib==2024.7
PyJWT==2.8.0
PyMsgBox==1.0.9
PyMySQL==1.1.1
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytweening==1.2.0
pytz==2024.1
pywin32-ctypes==0.2.2
requests==2.31.0
schedule==1.2.2
selenium==4.21.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.5.0
tqdm==4.66.4
trio==0.25.1
trio-websocket==0.11.1
twilio==9.2.3
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
webdriver-manager==4.0.1
Werkzeug==3.0.2
wsproto==1.2.0
yarl==1.9.4
zipp==3.18.1
